changed type from int to byte

+++ src/org/jgroups/protocols/FC.java
@@ -1,4 +1,4 @@
-// $Id: FC.java,v 1.17 2004/10/08 13:12:05 belaban Exp $
+// $Id: FC.java,v 1.18 2004/10/08 13:58:32 belaban Exp $
 
 package org.jgroups.protocols;
 
@@ -18,7 +18,7 @@ import java.util.*;
  * Note that this protocol must be located towards the top of the stack, or all down_threads from JChannel to this
  * protocol must be set to false ! This is in order to block JChannel.send()/JChannel.down().
  * @author Bela Ban
- * @version $Revision: 1.17 $
+ * @version $Revision: 1.18 $
  */
 public class FC extends Protocol {
 
@@ -451,14 +451,14 @@ public class FC extends Protocol {
 //    }
 
     public static class FcHeader extends Header implements Streamable {
-        public static final short REPLENISH = 1;
-        short  type = REPLENISH;
+        public static final byte REPLENISH = 1;
+        byte  type = REPLENISH;
 
         public FcHeader() {
 
         }
 
-        public FcHeader(short type) {
+        public FcHeader(byte type) {
             this.type=type;
         }
 
@@ -470,19 +470,19 @@ public class FC extends Protocol {
 
 
         public void writeExternal(ObjectOutput out) throws IOException {
-            out.writeShort(type);
+            out.writeByte(type);
         }
 
         public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
-            type=in.readShort();
+            type=in.readByte();
         }
 
         public void writeTo(DataOutputStream out) throws IOException {
-            out.writeShort(type);
+            out.writeByte(type);
         }
 
         public void readFrom(DataInputStream in) throws IOException, IllegalAccessException, InstantiationException {
-            type=in.readShort();
+            type=in.readByte();
         }
 
     }
